Double, Integer 동기화, 테스트 진행중

See merge request s08-bigdata-dist-sub2/S08P22C106!118

Former-commit-id: 9824179255d1207ab5269f16e0c7f67693626158

+++ back/src/main/java/bunsan/returnz/domain/game/dto/GameGamerStockDto.java
@@ -20,7 +20,7 @@ public class GameGamerStockDto {
 	private String companyCode;
 	private Integer totalCount; // 보유 수
 	private Integer totalAmount; // 총 주식 구매 가격
-	private Double averagePrice; // 구매 평균 단가
+	private Integer averagePrice; // 구매 평균 단가
 	private Double valuation; // 평가손익
 	private Double profitRate; // 수익률
 	private String logo;
